Refactor helper.sh and 01-control-plane-prepare.sh to improve variable handling and error management

+++ deploy/scripts/pipeline_scripts/v1/01-control-plane-prepare.sh
@@ -283,16 +283,12 @@ echo ""
 
 set -eu
 
-set -eu
-
 if [ -f "${deployer_environment_file_name}" ]; then
-
 	APPLICATION_CONFIGURATION_ID=$(grep -m1 "^APPLICATION_CONFIGURATION_ID" "${deployer_environment_file_name}" | awk -F'=' '{print $2}' | xargs || true)
 	export APPLICATION_CONFIGURATION_ID
 
 	DEPLOYER_KEYVAULT=$(grep -m1 "^DEPLOYER_KEYVAULT=" "${deployer_environment_file_name}" | awk -F'=' '{print $2}' | xargs || true)
 	echo "Deployer Key Vault:                  ${DEPLOYER_KEYVAULT}"
-
 fi
 echo -e "$green--- Adding deployment automation configuration to devops repository ---$reset"
 added=0
@@ -340,9 +336,10 @@ fi
 if [ 1 = $added ]; then
 	git config --global user.email "$BUILD_REQUESTEDFOREMAIL"
 	git config --global user.name "$BUILD_REQUESTEDFOR"
-	git commit -m "Added updates from Control Plane Deployment for $DEPLOYER_FOLDERNAME $LIBRARY_FOLDERNAME $BUILD_BUILDNUMBER [skip ci]"
-	if ! git -c http.extraheader="AUTHORIZATION: bearer $SYSTEM_ACCESSTOKEN" push --set-upstream origin "$BUILD_SOURCEBRANCHNAME" --force-with-lease; then
-		echo "##vso[task.logissue type=error]Failed to push changes to the repository."
+	if git commit -m "Added updates from Control Plane Deployment for $DEPLOYER_FOLDERNAME $LIBRARY_FOLDERNAME $BUILD_BUILDNUMBER [skip ci]"; then
+		if ! git -c http.extraheader="AUTHORIZATION: bearer $SYSTEM_ACCESSTOKEN" push --set-upstream origin "$BUILD_SOURCEBRANCHNAME" --force-with-lease; then
+			echo "##vso[task.logissue type=error]Failed to push changes to the repository."
+		fi
 	fi
 fi
